refactor(v1): move train-only wire serializers (tool_to_wire/serialize_completion) into TrainClient
They're only used by the renderer (build its generate request / set Response.raw), so they live
with the train client rather than the chat dialect. `message_to_wire` stays in the dialect — it's
also used by `extend` (user-sim over relay) and the default harness, and moving it would be a
circular import.
