+++ verifiers/v1/dialects/chat.py
@@ -98,9 +98,9 @@ def parse_tools(raw: list[dict] | None) -> list[Tool] | None:
 
 
 # --- vf -> chat wire ----------------------------------------------------------
-# The proxy relays the provider's raw response, so it never serializes; these are for the
-# renderer (which builds its generate request and has no raw to relay) and user-sim turn
-# injection. Public so the (chat-only) renderer can reuse them.
+# `message_to_wire` (chat-only): used by `extend` (user-sim turn injection), the default harness
+# (a Messages instruction), and the train client (its generate request). The proxy never
+# serializes — it relays the provider's raw bytes.
 
 
 def _content_to_wire(content):
@@ -136,57 +136,6 @@ def message_to_wire(message: Message) -> dict:
             "content": message.content,
         }
     return {"role": message.role, "content": _content_to_wire(message.content)}
-
-
-def tool_to_wire(tool: Tool) -> dict:
-    function: dict = {
-        "name": tool.name,
-        "description": tool.description,
-        "parameters": tool.parameters,
-    }
-    if tool.strict is not None:
-        function["strict"] = tool.strict
-    return {"type": "function", "function": function}
-
-
-def serialize_completion(response: Response, model: str) -> dict:
-    """A vf `Response` -> an OpenAI chat.completion dict the program's SDK expects. The renderer
-    sets this on `Response.raw` (it generates, so has no provider response to relay)."""
-    message: dict = {"role": "assistant", "content": response.message.content}
-    if response.message.reasoning_content is not None:
-        message["reasoning_content"] = response.message.reasoning_content
-    if response.message.tool_calls:
-        message["tool_calls"] = [
-            {
-                "id": c.id,
-                "type": "function",
-                "function": {"name": c.name, "arguments": c.arguments},
-            }
-            for c in response.message.tool_calls
-        ]
-    usage = (
-        {
-            "prompt_tokens": response.usage.prompt_tokens,
-            "completion_tokens": response.usage.completion_tokens,
-            "total_tokens": response.usage.total_tokens,
-        }
-        if response.usage
-        else None
-    )
-    return {
-        "id": response.id or "vf-intercept",
-        "object": "chat.completion",
-        "created": response.created,
-        "model": response.model or model,
-        "choices": [
-            {
-                "index": 0,
-                "message": message,
-                "finish_reason": response.finish_reason or "stop",
-            }
-        ],
-        "usage": usage,
-    }
 
 
 def response_from_wire(completion: ChatCompletion) -> Response:
